pyproject.toml
[project]
name = "messenger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firebase-admin==6.5.0",
    "flask==3.0.2",
    "flask-login==0.6.3",
    "flask-wtf==1.2.1",
    "pillow>=11.1.0",
    "python-dotenv>=1.1.0",
    "requests==2.31.0",
    "sqlalchemy==2.0.27",
    "werkzeug==3.0.1",
    "wtforms==3.1.2",
]
